requirements.txt
# requirements.txt
# Python 3.10+ required

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Geospatial processing
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# GTFS processing
gtfs_kit>=6.1.0
partridge>=1.1.1

# Census data
census>=0.8.21
us>=3.1.1

# Travel time / accessibility (optional - for job accessibility)
r5py>=0.1.1  # Requires Java 11+ and r5 jar

# Spatial statistics
pysal>=23.7
esda>=2.5.0
libpysal>=4.9.0
splot>=1.1.5

# Network analysis (for OSM data)
osmnx>=1.7.0
networkx>=3.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
contextily>=1.4.0
folium>=0.15.0

# Progress bars
tqdm>=4.66.0

# HTTP requests (for data download)
requests>=2.31.0
aiohttp>=3.9.0

# Jupyter (for notebooks)
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.26.0